speed up freecell moves BUG: 169245
svn path=/trunk/KDE/kdegames/kpat/; revision=862714

+++ dealer.cpp
@@ -45,7 +45,7 @@
 // ================================================================
 //                         class MoveHint
 
-MoveHint::MoveHint(Card *card, Pile *to, qint8 prio)
+MoveHint::MoveHint(Card *card, Pile *to, int prio)
 {
     m_card         = card;
     m_to           = to;
